Rename RenameVariableControl
